Refactor preprocessing stack into modular package
