+++ README.md
@@ -26,6 +26,10 @@ DiaRemot is a production-ready, CPU-only speech intelligence system that process
 - **AGENTS.md** - Setup guide for autonomous agents
 - **CLOUD_BUILD_GUIDE.md** - Cloud deployment instructions
 
+
+### Preprocessing architecture
+The preprocessing stack now lives under `src/diaremot/pipeline/preprocess/` with focused modules for configuration (`config.py`), disk I/O (`io.py`), chunk lifecycle management (`chunking.py`), denoising primitives (`denoise.py`), and the signal chain (`chain.py`). The legacy `audio_preprocessing.py` module remains as a façade so existing imports continue to work. A standalone CLI for manual runs is available at `scripts/preprocess_audio.py`.
+
 ---
 
 ## 11-Stage Processing Pipeline
