+++ src/diaremot.egg-info/SOURCES.txt
@@ -1,22 +1,17 @@
 README.md
 pyproject.toml
-src/diaremot/__init__.py
-src/diaremot/api.py
-src/diaremot/cli.py
+scripts/preprocess_audio.py
 src/diaremot.egg-info/PKG-INFO
 src/diaremot.egg-info/SOURCES.txt
 src/diaremot.egg-info/dependency_links.txt
 src/diaremot.egg-info/entry_points.txt
 src/diaremot.egg-info/requires.txt
 src/diaremot.egg-info/top_level.txt
+src/diaremot/__init__.py
 src/diaremot/affect/__init__.py
 src/diaremot/affect/emotion_analysis.py
 src/diaremot/affect/emotion_analyzer.py
 src/diaremot/affect/intent_defaults.py
-src/diaremot/affect/sed_panns.py
-src/diaremot/affect/sed_timeline.py
-src/diaremot/affect/ser_dpngtm.py
-src/diaremot/affect/ser_onnx.py
 src/diaremot/affect/paralinguistics/__init__.py
 src/diaremot/affect/paralinguistics/analysis.py
 src/diaremot/affect/paralinguistics/audio.py
@@ -28,6 +23,12 @@ src/diaremot/affect/paralinguistics/features.py
 src/diaremot/affect/paralinguistics/text.py
 src/diaremot/affect/paralinguistics/validation.py
 src/diaremot/affect/paralinguistics/voice.py
+src/diaremot/affect/sed_panns.py
+src/diaremot/affect/sed_timeline.py
+src/diaremot/affect/ser_dpngtm.py
+src/diaremot/affect/ser_onnx.py
+src/diaremot/api.py
+src/diaremot/cli.py
 src/diaremot/io/__init__.py
 src/diaremot/io/download_utils.py
 src/diaremot/io/onnx_runtime_guard.py
@@ -40,29 +41,14 @@ src/diaremot/pipeline/auto_tuner.py
 src/diaremot/pipeline/cache_env.py
 src/diaremot/pipeline/cli_entry.py
 src/diaremot/pipeline/config.py
-src/diaremot/pipeline/cpu_optimized_diarizer.py
-src/diaremot/pipeline/dependency_checks.py
-src/diaremot/pipeline/diagnostics_smoke.py
-src/diaremot/pipeline/errors.py
-src/diaremot/pipeline/logging_utils.py
-src/diaremot/pipeline/orchestrator.py
-src/diaremot/pipeline/outputs.py
-src/diaremot/pipeline/pipeline_checkpoint_system.py
-src/diaremot/pipeline/pipeline_config.py
-src/diaremot/pipeline/pipeline_diagnostic.py
-src/diaremot/pipeline/pipeline_healthcheck.py
-src/diaremot/pipeline/provider_registry.py
-src/diaremot/pipeline/run_pipeline.py
-src/diaremot/pipeline/runtime_env.py
-src/diaremot/pipeline/speaker_diarization.py
-src/diaremot/pipeline/speaker_diarization_batched.py
-src/diaremot/pipeline/transcription_module.py
-src/diaremot/pipeline/validate_system_complete.py
 src/diaremot/pipeline/core/__init__.py
 src/diaremot/pipeline/core/affect_mixin.py
 src/diaremot/pipeline/core/component_factory.py
 src/diaremot/pipeline/core/output_mixin.py
 src/diaremot/pipeline/core/paralinguistics_mixin.py
+src/diaremot/pipeline/cpu_optimized_diarizer.py
+src/diaremot/pipeline/dependency_checks.py
+src/diaremot/pipeline/diagnostics_smoke.py
 src/diaremot/pipeline/diarization/__init__.py
 src/diaremot/pipeline/diarization/clustering.py
 src/diaremot/pipeline/diarization/config.py
@@ -74,10 +60,29 @@ src/diaremot/pipeline/diarization/registry.py
 src/diaremot/pipeline/diarization/segments.py
 src/diaremot/pipeline/diarization/utils.py
 src/diaremot/pipeline/diarization/vad.py
+src/diaremot/pipeline/errors.py
+src/diaremot/pipeline/logging_utils.py
+src/diaremot/pipeline/orchestrator.py
+src/diaremot/pipeline/outputs.py
+src/diaremot/pipeline/pipeline_checkpoint_system.py
+src/diaremot/pipeline/pipeline_config.py
+src/diaremot/pipeline/pipeline_diagnostic.py
+src/diaremot/pipeline/pipeline_healthcheck.py
+src/diaremot/pipeline/preprocess/__init__.py
+src/diaremot/pipeline/preprocess/chain.py
+src/diaremot/pipeline/preprocess/chunking.py
+src/diaremot/pipeline/preprocess/config.py
+src/diaremot/pipeline/preprocess/denoise.py
+src/diaremot/pipeline/preprocess/io.py
+src/diaremot/pipeline/provider_registry.py
+src/diaremot/pipeline/run_pipeline.py
 src/diaremot/pipeline/runtime/__init__.py
 src/diaremot/pipeline/runtime/environment.py
 src/diaremot/pipeline/runtime/executor.py
 src/diaremot/pipeline/runtime/session.py
+src/diaremot/pipeline/runtime_env.py
+src/diaremot/pipeline/speaker_diarization.py
+src/diaremot/pipeline/speaker_diarization_batched.py
 src/diaremot/pipeline/stages/__init__.py
 src/diaremot/pipeline/stages/affect.py
 src/diaremot/pipeline/stages/asr.py
@@ -89,6 +94,8 @@ src/diaremot/pipeline/stages/paralinguistics.py
 src/diaremot/pipeline/stages/preprocess.py
 src/diaremot/pipeline/stages/summaries.py
 src/diaremot/pipeline/stages/utils.py
+src/diaremot/pipeline/transcription_module.py
+src/diaremot/pipeline/validate_system_complete.py
 src/diaremot/sed/__init__.py
 src/diaremot/sed/sed_panns_onnx.py
 src/diaremot/sed/sed_yamnet_tf.py
@@ -106,3 +113,5 @@ src/diaremot2_on/__init__.py
 src/diaremot2_on/diarization.py
 tests/test_diarization.py
 tests/test_outputs_transcript.py
+tests/test_paralinguistics.py
+tests/test_preprocess_stack.py
